deploy/ddos-protection/requirements.txt
# DDoS Protection Service Dependencies
aiohttp==3.9.3
prometheus-client==0.19.0
pyyaml==6.0.1
redis==5.0.1
asyncio-throttle==1.0.2
python-dateutil==2.8.2
structlog==24.1.0